Make `onchange` mandatory.
Since a select with no associate action is useless this takes the component
one step further towards explicitness.

Now the component won't do ANYTHING AT ALL when an element is selected, it will relay
completely in the `onchange` action defined by the user. The user has to define that
action and update the `selected` value to propagate the change to the component again.