Fix placeholder in multiple select for good.
Tests were passing in phantomjs but the feature was actually broken in real life.
Apparently there is some kind if irregular on how attributes are treated when it's value
is undefined.
Now the placeholder is always present, but when falsey its value is an empty strings. That
behaves consistently across browsers.