Refactor components composition
The way compositions was handled was a bit chaotic. That was an internal issue but
still an issue. Some actions received the `dropdown` and other didn't, and
you had to remember to pass the dropdown as last argument to function calls.

This refactor creates a uniform shape for the yielded public API of the component
as follows:

```js
{
  state1: 'foo',
  state2: 'bar',
  ...
  actions: {
    action1: fn,
    action2: fn,
    ...
  }
}
```

Each level can use the public API of the nested component as follows:

```hsb
{{!-- in delorean.hbs --}}
{{#flux-condensor as |condensor|}}
  {{#with (hash gammaPower=condensor.level actions=(hash goToFuture=(action "travelForward" condensor) goToPast=(action "travelBackwards" condensor))) as |delorean|}}
    {{control-panel delorean=delorean year=currentYear}}
  {{/with}}
{{/flux-condensor}}

{{!-- in control-panel.hbs --}}
Year: {{eight-segments-display number=year}}
Charge: {{eight-segments-display number=delorean.gammaPower}}
<button onclick={{action delorean.goToFuture 10 'years'}}>+10</button>
<button onclick={{action delorean.actions.goToPast 10 'years'}}>-10</button>
```

The division between state and actions is artifical but help to get my head around problems.

What goes inside this publicAPI and what is passed as separate attributes to the `selected` or `options`
components? Basically only the things I feel confortable considering public API. Eventually this "publicAPI-object"
might be the only passed attribute.