Update ember-basic-dropdown to fix rendering issue in ember 2.2.0+
The bug was always there but ember 2.2.0 triggered it, perhaps for being
faster or something.

tl;dr; Ember-basic-dropdown has to ensure that the opened dropdown is
closed before the new one opens (otherwise the reposition code will target
the first one).
The solution is to attach the click event to close the dropdown when clicked outside
in the capture fase, instead of in the bubbling fase.