Some changes for refactoring the internals with one goal: Dogfooding
Ember-power-select should be flexible enough to be able to build the multiple
select on top of itself. I've realized that the primitives that the component exposes
are quite flexible but not there yet.

- The `select(dropdown, newSelection, e)` action becomes more naive. It just checks if the element being selected
is different from the current selection, and if so, it invokes the `onchange` action.
The selection can be an array (in multiple selects is). It doesn't care.
- A new `choose(dropdown, option, e)` action is the one that selects a particular option. In
single mode is basically an alias of select, but in multiple mode is generates the new selection
from the existing one by cloning it and adding or removing the choosen option from it. At the
moment this method lives in the multiple select but it should live in the `options` component.