Simplify logic of the cursor position.
Every time the `searchText` is updated, the component containing the input
has to propagate that change. For propagate the change and at the same time
preserve the cursor position, the simplest strategy is to compate the value
of the input before updating it.

If the value about to be updated is the same that the input has, we can skip the
update. That will prevent to loose the cursor position while typing in the middle
of the text.

However, updates in which the value being set is different from the value in
the input will wipe the cursor position, as expected.