Completely control focus from within the beforeOptionsComponent
Ember-basic-dropdown no longer call preventDefault on mousedown events
on the trigger. That means that the trigger will get the focus normally.
The beforeOptionsComponent used to docus the input inside when rendered.

The problem was that the trigger gets the focus after the component has
been rendered and steals the focus back.

With this change, the beforeOptionsComponent receives an event when the trigger
is focused, and can then safely set the focus to the input.