Add FastClick compatibility
Add FastClick compatibility

Fastclick does not seem to properly handle mousedown events. Adding the onclick event allows the addon to be used on mobile devices in combination with FastClick.

Partly solves issue #248 described in ember-power-select "Incompatibility with fastclick?".
https://github.com/cibernox/ember-power-select/issues/248

Depends on https://github.com/cibernox/ember-basic-dropdown/pull/53 to be accepted in ember-basic-dropdown. Otherwise the dropdown won't even open to allow for selecting an option.