Improve docs pages scroll handling
This will cause every public page to start out scrolled to the top, and then on when returning to a visited route will return to the last remember scroll position.

This makes moving forward and backward through the documentation pages feel nicer, because forward links scroll you to the top automatically, and backward links take you back to where you were when you clicked forward.