[BREAKING] e.preventDefault no longer prevents the default component's behaviour (#449)
Before users could use `return false` or `preventDefault`. Now only `return false` works.

The reason for that is that some events need to be `defaultPrevented` to avoid browsers'
behaviour (like touchEvents to avoid trigger synthetic clicks), and sometimes preventing
the default behaviour of a event has consecuences we don't want (like preventing keydowns
make the input receiving the keydown to not update it's value).

For that, the idiomatic way now of preventing the default behavior is return false
from the `on*` actions.