[BUGFIX] Decouple scrolling from highlighting
That way highlight with the mouse doesn't trigger a scroll change, and
at the same time allows to scroll reliably using `didInsertElement` that
we have the guarantee that is called once the dom is ready.

This adds a nice new `scrollTo(option, dropdown, e)` action to the public API object received
by all the hooks of the component that can be used to command the select to scroll
to any option.