Add support for cancellables in search action.
With ember-concurrency gaining popularity every day and cancellable promises
around the corner (sic) it's time to start being good citizens and cancel all
work that can be cancelled when the component is no longer interested in it.

There is a couple situations where this can happen:

If the `search` action returns a `thenable` object that responds to the
`.cancel()` method, firing another search before the first is completed will
cancel the previous one. This was possible before by just adding `.restartable()`
to your e-c task, but now it will work with any cancellable object.

Another situation where this could happen is when, after firing a search that
returns can cancellable-thenable, and before that search completes, the searchbox
gets cleared. In EPS, clearing the search doesn't fire another search, so this
situation was not easy to detect and cancel properly with a `.restatable()` task.
Now, EPS detects this situation and cancels the search.

Closes #693