[FEATURE] Allow to customize what element is highlighted by default
By default when a select component opens, the highlighted element is the selected one,
or the first option in the list if nothing is selected. This is the right behaviour for
99% percent of the cases.

However, there might be some situations where the user wants to customize this, and now
it's possible passing `defaultHighlighted=something`.

That option can be either one of the options, or a function that takes the publicAPI
of the component and returns one of the options.