update `contentId` during async tests
I've got a test that's failing b/c the power select is only rendered when the promise is fulfilled. Due to the async op, `.ember-power-select-trigger` isn't rendered, so `$trigger` is defined in the fallback clause. The element that's defined for `$trigger` is outside the scope of the `power-select` component, and so has no `aria-owns` attribute.

This patch simply runs through the "find a trigger" logic once again in the `andThen` block if `contentId` is undefined.

also I put the "find a trigger" logic in a separate function.