Ignore keydown with modifier keys pressed
If we do not ignore them, we end up, for instance, filtering by "r" when when we
press ctrl+r to reload the page, or filtering by "l" when doing ctrl+l to select
the browser's URL.

It is always a bit annoying to witness this involuntary filtering, but more so
when the widget auto-saves its content. Especially so with ctrl+r, with which
you might not even notice the change because of the page reload.