Enable native-like search logic
**NOTE 1** This PR does not changes the default matcher. For a full,
native-like experience, a matcher should be provided (see end of this
commit).

**NOTE 2** This PR does not change anything about diacritics. Any time
a match is done, it is understood that the diacritics are normalized
(case folding).

**NOTE 3** For this behaviour to work when the select is open, the user
needs to disable the search.

About this PR
======

This PR aims to give EPS a more native-like behaviour. A complete change
would heavily break the behaviour of EPS.

What's the UX in current browsers?
======

For this, we compared safari, chrome, and firefox. Both safari and
chrome have different behaviours depending on whether the select is
open or closed. Firefox does not. We believe the select should behave
the same in both states, and a deeper look at blink's code seem to proof
it. For the purposes of this PR, the behaviour explained and implemented
is the one reproducible when the select is closed in safari and chrome
(firefox only implement this behaviour).

When a user focuses on a select, does not matter the state, they can
go to the next option that matches the treated input. The value the user
enters does not need to be the one that is used for matching, but
whatever the user types, it might not be what it's used for looking for
matches. The system keeps its 1s memory before forgetting the previous
search term (as it does now). Furthermore, the threshold in EPS is the
same we can find in webkit and blink.

The differences that this PR implements are two: search offsets and
cycling through the results. These behaviour are explained in depth
next.

Search offset
======

The original implementation applies an offset when looking for matching
results. If we are in _cycling mode_ (see next part) or starting a new
search (by focusing or not typing for 1s), the select will start looking
one result after the current highlighted / selected option. That means
that if we have selected an option that matches our current intention,
the user will not stay there, but go to the next option (if any). Once
you think about it, seems natural that if the user is interacting with
the component, it is because it is not interested in the current
selection.

On the other hand, if the user is typing a longer search term, the
offset can be 0, since the current option might be a valid one (changing
the original selection was already done when typing the first character,
see paragraph above).

In any case, current highlighted / selection is added to the offset
above.

Cycling through the results
======

When a user focuses on a select and types a letter, let's say _a_, the
select will show the first result that matches _a_. If the user, within
the time threshold, types _a_ again, the select will go to the next
result that matches _a_ and not _aa_. That way, typing again the same
first letter several times let us cycle through the results.

It is important to note that this only works for the first character.
For example, _Unittt_ will not cycle through the result for _Unit_ in
a country selection.

Also, the repetitions need to be remembered. In a select full of names,
typing _aar_ will take us to the right selection (_Aaron_) and not to
a result for _ar_.

Show me the (original) code!
======

> Files linked here are pinned to the latest commit they were changed in.

The inital research was done via manual testing. Then, after inspecting
[blink's HTMLSelectElement
implementation](https://github.com/ChromiumWebApps/blink/blob/f26b5a9078d12d954a236c557776f4b1395726c6/Source/core/html/HTMLSelectElement.cpp), I studied
[blink's
TypeAhead](https://github.com/ChromiumWebApps/blink/blob/b6720b1767fb5208aa421237c087f062093929e7/Source/core/html/forms/TypeAhead.cpp). Current implementation is thought as a mix of
[TypeAhead::handleEvent](https://github.com/ChromiumWebApps/blink/blob/b6720b1767fb5208aa421237c087f062093929e7/Source/core/html/forms/TypeAhead.cpp#L60)
and
[filterOptions](https://github.com/cibernox/ember-power-select/blob/6b1e472d394a9322f9a25a668cf0686bbeb2ad2c/addon/utils/group-utils.js#L74-L95).

I stated above that the lack of cycling behaviour when the select is
open is quite likely a bug because of
[this comment that can be found in blink's
PopupListBox](https://github.com/ChromiumWebApps/blink/blob/f26b5a9078d12d954a236c557776f4b1395726c6/Source/web/PopupListBox.cpp#L320-L321).

Native matcher
======

The native matcher behaviour is just to look for a prefix.

```js
import { stripDiacritics } from 'ember-power-select/utils/group-utils';

function nativeMatcher(value, text) {
  return stripDiacritics(value).toUpperCase().startsWith(stripDiacritics(text).toUpperCase()) ? 1 : -1;
}
```

Live example
======

You can find the behaviour of this PR in http://power-test.surge.sh/

- Try looking for _Aaron_
- Cycling behaviour is better done with the "F" (shorter and the group
in the second select is repeated).