Make `results` and `resultCount` computed properties
If those properties are just getters, they get called pretty often: `results`
was called 4 times for every keystroke in the search box, and `resultsCount` 
twice.

Since filtering iterates over every element of the list (potentially a hundreds 
of elements), calling it so ofter was possibly inefficient. 
Counting the results also is a function that iterates all options (and suboptions 
if there are groups) so it better to cache the result.

From a first swipe of the component, those two seem the only getters worth caching