Refactor I'm done with you joke
The need to refactor came from no-setter-return eslint rule.

This refactoring slightly changes the behavior:

Before the refactoring:
1. User reaches the "I'm done with you" step, the dropdown is disabled.
2. User navigates away and goes back to 'public-pages.index',
   the dropdown is enabled again on the step "I'm done with you".

After the refactoring:
On #2 the dropdown remains disabled.

In order to keep the original behavior it is needed to listen to the
router's 'routeDidChange' event.

But I'm not sure if it was an intended behavior in the first place.