[bugfix] remove remaining observer on PowerSelect teardown
PowerSelect Component instances accumulate in the JS Heap when the
initial 'selected' value is a PromiseProxy. That's because we
addObserver, but the observer is not removed prior to component
teardown.

We now remove the observer on component teardown, which allows
PowerSelect Component instances to be removed from the JS Heap when
the Component is destroyed.

Fixes #1391