[bugfix] remove observer on 'selected' update
PowerSelect Component instances accumulate in the JS Heap when the
upeated 'selected' value is ever PromiseProxy. That's because we
addObserver, but the observer is not alwaysn removed if switching
between a PromiseProxy value and a non-PromiseProxy value.

We now remove the observer on 'selected' update, which allows
PowerSelect Component instances to be removed from the JS Heap when
the Component is destroyed.