[bugfix] onChange action receives Event as last arg on mouseup
This change ensures that when an option in the PowerSelect/Options
component is clicked, an action registered on mouseup will pass the
Mouseup Event all the way to the onChoose action.

Previously, the second AND third event passed to the onChoose action
was `this.storedAPI` due to arguments not being passed as expected,
and the Event was never received by the `onChoose` action.