Cleans up scrollTo's element finding
This commit replaces a somewhat convoluted query chain with
a single querySelector to fetch the specific element we want to
focus on.

This feels slightly cleaner to my eyes, but has the added benefit
of ensuring it works correctly with something like "vertical-collection",
which occluding certain elements from the list for performance reasons.
This was causing keyboard-navigation to fail unexpectedly, since `.item(19)`
was no longer guaranteed to be the item with `data-option-index='19'`, or
even to exist.