fix: Remove `ember-cli-deprecation-workflow`
There's [an issue with ember-cli-deprecation-workflow](https://github.com/mixonic/ember-cli-deprecation-workflow/issues/133)
that causes an error during CI for the embroider optimized build.

I think we can remove `ember-cli-deprecation-workflow` from this addon
because it does not seem to be doing anything useful but maybe I'm
missing something?

This was originally added in
https://github.com/cibernox/ember-power-select/commit/c10dd8eb5dc89a439541dac08fedd7878f9d544d
as part of https://github.com/cibernox/ember-power-select/pull/1431 but
there is not really much context to go off of as far as why it was
added. It seems like maybe it was added while debugging GitHub Actions
and accidentally left laying around?

- https://github.com/mixonic/ember-cli-deprecation-workflow/issues/133
- https://github.com/embroider-build/embroider/tree/main/packages/test-setup#embroideroptimizedembertryconfig