fix: Remove ember-cli-fastboot
This removes `ember-cli-fastboot` because it is not currently compatible
with embroider optimized builds.

I believe `ember-cli-fastboot` was added as a dependency so
`ember-power-select` could run some tests to ensure compatibility. This
makes sense but AFAICT the tests to check fastboot compatibility have
not actually been running since 2016. Since these fastboot tests have not
actually been running for so long, IMO it makes sense to remove all of
this for now and revisit fastboot tests separately.

- Stopped running fastboot tests https://github.com/cibernox/ember-power-select/commit/1b9f90d123e95e4abfc72974946eb609a2149156#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L13