Migrate test helpers and some tests to TypeScript (#1743)
* Convert test helpers to TypeScript

* Replace `npm` with `pnpm`

* Update actions

* Remove duplicated file

There is a TS version of that file

* Convert some tests to TypeScript

To illustrate that the test helpers have correct types.

`clearSelected` doesn't have 2nd arg. Fix it in the docs and in the tests.

Fix path to helpers `setup*`, we should local versions instead of
ember-qunit ones.

Fix type of the argument of stripDiacritics, as it accepts numbers as well.