Use Promise instead of PromiseProxy
A couple of things were typed as `PromiseProxy`s when they only used the `Promise` interface. This prevented callers from, for example, supplying a search function that produces a regular promise, i.e. a plain old async function. So change them to `Promise`.

The `@selected` argument still uses a `PromiseProxy` because the component observes its `content` property, so allowing it to be a `Promise` would be a more involved code change.