Type improvements (#1746)
* Make array arguments immutable

Type the array component arguments as immutable. The component doesn't actually mutate them, so typing them as readonly allows caller to pass in either mutable or immutable arrays, rather than only mutable ones.

* Use Promise instead of PromiseProxy

A couple of things were typed as `PromiseProxy`s when they only used the `Promise` interface. This prevented callers from, for example, supplying a search function that produces a regular promise, i.e. a plain old async function. So change them to `Promise`.

The `@selected` argument still uses a `PromiseProxy` because the component observes its `content` property, so allowing it to be a `Promise` would be a more involved code change.